Add files via upload
Code for find the PTM potentiality rate
To determine the PTM potentiality rate (PPR) of metabolic enzymes, a method that integrates amino acid composition with observed PTM occurrences was employed. For each enzyme, the total count of each amino acid type in its sequence was calculated to estimate the potential number of modification sites for each PTM, based on the amino acid's susceptibility to a specific modification (e.g., lysine for acetylation, serine/threonine/tyrosine for phosphorylation). The actual count of PTM occurrences was obtained from the dataset, representing experimentally validated modification events. PPRs were then calculated as the ratio of the actual count to the potential count, expressed as a percentage. This approach allowed us to quantitatively assess the propensity of metabolic enzymes to acquire PTMs, while normalizing for variations in amino acid composition and sequence length.