Add files via upload
To find the cumilative frequency for each PTM site
To classify PTM sites that are predominant within the specific PTM types in metabolic enzymes, a cumulative frequency analysis was performed. The frequency of detection of each PTM sites for every metabolic enzyme was calculated based on the number of research articles that reported them. The cumulative percentage was calculated by summing the reference counts for each PTM site sequentially and dividing it by the total reference count for that modification type in the enzyme.